\name{cardinality}
\alias{cardinality}
\title{Cardinality of Column Vectors}
\usage{
  cardinality(W)
}
\arguments{
  \item{W}{a numeric matrix, e.g. the rotation matrix of a
  sparse PCA analysis}
}
\description{
  Computes the cardinality (the sum of non-zero elements)
  of each column of the matrix \eqn{\mathbf{W}}{W}.
}

